Don't throw exceptions if we already had the file during a rename
If we rename a file from Foo.cs to Foocs, we went through
OnSourceFileRenaming and updated it, but we won't get a call if we do
that rename back, since we are tracking renames to documents in our
projects and the second rename doesn't "count". This could result
in you having both names, and if you renamed it again bad things would
happen.