Fix capturing of a generic type in local functions
This changes our generic-lowering logic to use 'Arity' instead of
'IsGenericType' to figure out whether or not a local function frame
needs generic subsitution. We previously used IsGenericType, which is
inappropriate because it also considers whether or not the containing
type of the frame is generic. The containing type is necessarily already
subsituted when calling a local function, so it will never be possible
to substite type parameters.

Instead, we should use Arity of the frame, which is only non-zero if the
frame itself is generic, regardless of the containing type.