Use the RC SDK to build if you're on RC
This allows build to succeed at installing the extensions, so that it's
possible to debug changes.

The wrinkles here are that:
1. SDK's before the RC one don't know how to install on RC.
2. The RC SDK *requires* a <Prerequisites> element
3. The Willow installer fails to install vsixes that have a <Prerequisites> element.

To workaround 2 and 3, if we detect that you are building with something before RC,
we run an xsl transform to remove the <Prerequisites> element before the SDK uses it.

Since our official builds for insertion into Willow are still on VS 2015, they won't
have the element.