Fix and verify Microsoft.Net.Compilers.nuspec
The NuSpec file for the compiler is missing a number of facades.  This flaw's been around for some time in a few forms but it's gone un-noticed due to our reflection magic around facades.  Now that the magic is gone we're finding places where the NuSpec is incorrect.

This change both fixes the NuSpec and adds verification that it is correct.  The strategy for correctness is to build a deployment project which contains all of the deliverables in the NuSpec.  Namely:

- csc.exe
- vbc.exe
- csi.exe
- Microsoft.Build.Tasks.CodeAnalysis.dll

The MSBuild process will deploy all facades necessary for execution to the output directory.  That makes it possible to do a simple check that the output and NuSpec are consistent with each other.