Migrate Roslyn to build on Preview5
Migrate most of Roslyn to build on top of Dev15 Preview5.

1. Update project.json files to reference the Dev15 Preview5 version of
packages. This generally replaces the Dev14 Update3 versions.
2. Update XAML files to refer to the 15.0 version of assemblies, rather than
the 14.0 version.
3. Eliminate uses of the IErrorList2 interface and related types. This
basically controlled whether or not we would run whole-solution analysis, but
we now approach the problem in a different way. In Dev15 this interface is no
longer defined or implemented, so we can just delete all the code that depended
on it.
4. Replace some uses of the Dev15BinariesForRoslyn package. This basically
wrapped up a bunch of other VS binaries that are now available in their own
official packages; as such this is no longer needed.

This does not update the expression evaluators for the debugger, as those are
being handled by @cston in https://github.com/dotnet/roslyn/pull/14858.