Build story for portable unit tests
Portable unit tests are difficult because as portable libraries they can't truly be built in a runnable fashion.  Instead they need a deployment project to provide a runtime on which to delpoy the output of the build.  To support running our portable tests on both desktop and CoreClr the build is structured as follows:

- Portable tests build w/o deploy to a Dll directory
- Desktop delpoyment references all portable tests and outputs to UnitTests\Desktop\Portable
- CoreClr deployment references all portable tests and outputs to CoreClrTests

This breaks F5 for the moment but gives us the flexibility to restore later.