Don't crash during shutdown if we try creating metadata snapshots
Windows Error Reporting is insisting it's surprisingly common for us
to have a thread trying to create a compilation while Visual Studio
is shutting down and we can't make COM calls anymore. This prevents
that from happening. This is a tactical fix and we'll still let the
rest of metadata snapshot system work normally.