Update MSBuildWorkspace to find MSBuild binaries
To use `MSBuildWorkspace` in an app currently requires that the MSBuild
binaries be GAC'd, or that you basically have an entire MSBuild install in
the binding path of your app. Since the MSBuild binaries are no longer
GAC'd starting with Dev15, this poses a problem for developers. Even our
`MSBuildWorkspace` tests (and other unit tests that happen to use it) run
into this issue as having a few loose MSBuild binaries next to the unit
test assembly isn't enough--you need a full MSBuild install.

This commit makes two changes. First, the `MSBuildWorkspace` factory
methods have been overloaded to take the full path to an MSBuild install
as a parameter. If supplied, a basic `AppDomain.AssemblyResolve` handler
takes care of loading the MSBuild assemblies from this location as
needed. If not supplied, a developer will need to supply their own
mechanism for loading the assemblies.

Second, our `MSBuildWorkspace` unit tests have been updated to make use
of this new functionality by finding an MSBuild installed by Dev15 and
passing its path.