Further updates to app.config and project.json files
These are further changes needed to get unit tests working on a system
with only Dev15 Preview5. Notably, the latest version of
Microsoft.VisualStudio.Composition has an undeclared dependency on
System.Threading.Tasks.Dataflow. This requires that we include it in a
bunch of unit tests projects. Further, the dependency is on version 4.5.24,
but that version isn't available on NuGet. Instead, we use 4.5.25 and
introduce binding redirects where needed.