Only put Git SHA in official build binaries
Putting the GIT SHA into unofficial binaries was hurting our efforts around determinism.  It meant that binaries authored by developers which had the same code, but against different HEAD values, would end up with distinct binaries.  Hence anything in our infrastructure which leveraged determinism for caching wasn't getting the benefits it should.

The primary use of the GIT SHA is to help us track back from a released binary to the source that created it.  Especially important given the frequency with which we produce NuGet packages.  Hence it only needs to be present on official built binaries as they are the only ones we release.  Having it into the developer builds doesn't have any notable value.