Remove the depedency projects
The dependency projects were added as a way to help unify our project.json references.  Instead of every project referencing the NuGet package, we had a dependency project which referenced it and everyone referenced that dependency project.  Due to the transitive nature of NuGet it would be roughly equivalent.

This system had a number of downsides:

1. There was no enforcement.  Even though dependency projects existed, there was no mechanism to force developers to use them.
2. Getting the granularity correct was tricky.
3. The projects themselves have special props / targets to get them to produce no output.  This can trip up MSBuild up to date checks.

Since then though we've begun using RepoUtil to verify our NuGet reference correctness.  This removed the need for these projects and hence I'm removing them from our build.