Make BatchFix code action application deterministic.
Introduce a new, deterministic BatchFix application algorithm. The new algorithm works in the following manner:

When producing code-actions, we keep track of both the code-action and the diagnostic that was used to create it. When we later go and 'apply' the code actions in a single document, we'll process them in diagnostic-location order. Is is as if we took each diagnostic of a document in order, produced the CodeAction for that diagnostic, and then applied that CodeAction.

I also took this opportunity to clean up and parallelize much of this code. We now compute code actions in parallel, and we process all documents in parallel when applying changes. Once all application is done, we can then take all the final results and apply them to the solution.