Add a portable ThrowingTraceListener
The throwing version of a TraceListener is how we prevent Debug.Assert dialogs popping up on our Jenkins runs.  The separation of our tests into a proper portable and non-portable subset revealed that we only had a desktop version of this class.  A portable version is needed to get the same behavior for CoreCLR unit tests.