Ignore use-before-declaration in local functions (#15339)
Prior to C# 7 we tried to avoid giving duplicate errors
when the user used a variable before it was declared (which
is both a use-before-declared and a use-before-assigned error),
but the line here is blurred with local functions, which can
use variables before they are declared in subtle ways that can't
be detected based on declaration order in syntax.

It may be possible to provide the earlier behavior, but the
current optimization is definitely insufficient and hides
legitimate errors in user programs.

Fixes #15298
Fixes #15322

(cherry picked from commit 1c8372cea14d7b75e8b55c6c38314291af165d5f)