Fix issue with IVsNavInfo for project in solution folder
Fixes #14462

In the VS 2015 Update 1 timeframe, I introduced a bug when moving from new'ing up a StringBuilder to using a pooled StringBuilder. When it was new'ed up, the StringBuilder was initialized with a value. When moving to a pooled StringBuilder, I forgot to initialize it with this value. This has the effect of producing an invalid IVsNavInfo for items in solution folders.