Work around unexpected synchronous blocking in FileStream.ReadAsync()
We've seen a performance issue where the user gets a temporary hang
and a symptom of this is an excessive number of blocked thread pool
threads on IO. This was because synchronous work is assuming another
thread pool thread is available to complete the work. By tweaking how
we do the reads, we can work around this issue.

Fixes https://connect.microsoft.com/VisualStudio/feedback/details/2087450/visual-studio-hangs-on-build-save-and-start-for-xaml-projects-microsoft-visual-studio-is-busy