Add --skipcrossgen to cibuild.sh
In the general case, finding the correct crossgen to invoke is non
trivial since there's no strong corespondence between the host Runtime
ID (which is easy to compute) and the RID that crossgen would be
packaged in.

For example both RHEL and Centos have crossgen packaged in a nuget
packages that uses `rhel.7-x64` as the RID, in order to share the
package across minor versions of RHEL and Centos (which is binary
compatibile).

Since we can build Roslyn without crossgening the bootstrapped
toolchain, I'm exposing an option to allow skipping of crossgen for
platforms where a developer knows apriori that `crossgen.sh` would not
compute the correct RID and fail.