Fix caching logic
The caching logic is now much stricture about the handling of missing
references during caching.  There are valid cases for missing
references, light up features.  But there are other cases where it
simply repreesnts an error.  The code now defaults to assuming missing
references are an error and opts in the cases where it's known to be
okay.