Return to capturing the foreground thread during ForegroundThreadAffinitizedObject creation
In 44142e5fee9 I removed this code, under the belief that it wasn't
a problem. However, when WpfFact-attributed tests run, we clear the
foreground thread. Some static state (namely the
ForegroundNotificationService) is still trying to operate between test
runs, so it'd see the fallback state and then bad things would happen.
Although there's a number of terrifying things there, for now we'll
just roll it back.