Don't capture what the foreground thread might be until needed
When we create a ForegroundThreadAffinitizedObject, it captures
what is believed to be the foreground thread at that point. That can be
bad if it hasn't been initialized, as the background thread might be
the assumed initialization. As a terrible, slimy hack, we'll just defer
this until the foreground asserts are happening. The first might slip
through but later requests would then blow up.