Allow closures to take struct- and class- environments
Lambda rewriting currently allows frames to be structs in the instance
where

  1. The variables in the frame are not captured by any closures
  converted to delegates.

  2. A closure nested inside the struct closure captures a separate
  variable that does require a class frame.

This poses a problem when the "outer" closure is in the same frame as
the variable that requires a class closure. The problem is that we treat
scopes and frames as essentially the same thing -- if any scope is
perceived to require a particular environment type, all captured
variables in that scope will be captured on that frame and all closures
will be lowered onto that frame.

This creates a conflict between class-based environment capturing, where
we want to capture the frame pointer as a field, and struct-based frame
capturing, where we want to add arguments as ref parameters. Doing both
results in breaking the delegate contract (no signature rewriting) or
losing required struct arguments to intermediate closures.

This change adds struct-based environments even to closures in
class-frames, as long as they are not converted to delegates, async, or
iterator functions. Otherwise, all captured environments are forced to
be class-based.

Fixes #16895