CodeCleanup should not try adding "()" to tuple element names.
Fixes  #16781

Becasue of the another bug, #16697, it is possible that tuple element names would bind to methods. That will require a bigger fix.

But regrdless of that, the CodeCleanup should not try adding "()" to tuple names. No need to consult with semantic info (which is fairly expensive), when this will never be a right thing to do. It does not fit the syntax model and crashes with invalid cast exception.