Fix SolutionExplorer_InProc.OpenFile
Currently this method attempts to look up an `EnvDTE.Project` by name, specifically by passing in the project name to the `EnvDTE.Projects.Item(...)` method. As useful as it would be to look up a project by name, the `Item` method doesn't actually accept a string as an argument. Instead, we need to iterate over all the projects (starting at index 1, of course) and compare the `Name` property.