Take a screenshot when an Assert fails
Take a screenshot whenever an xUnit `Assert` method fails. We achieve this by having the `VisualStudioInstanceFactory` hook the `AppDomain.FirstChanceException` event. If the thrown exception is an `XunitException` we capture the screen using the `ScreenshotService` and save it to a time-stamped file, like "xUnitResults\Screenshots\EqualException-13.19.52.png". That is, the screenshot was triggered by the failure of an `Assert.Equal` method at 1:22 PM (and change).

We need to be careful to catch and handle all exceptions that might arise in our `FirstChanceException` handler. In this case we just write it to the console and swallow it.