Ensure proper tracking/cleanup of _bufferToDocumentInCurrentContextMap when closing shared documents
Current implementation was leaking entries in this map for shared document buffers when the shared document was closed. The primary reason was that we were only tracking the active document ID corresponding to a given text container. Hence, on document close for a specific document ID, we could not determine if we still need to track the text container for other related document IDs or not. This leak in turn was causing stale entries in the map to be returned when the shared document was re-opened.

Fix is to track all the related documents that have been opened with a specific text container, and remove the textContainer tracking entry in _bufferToDocumentInCurrentContextMap when all the attached document IDs have been closed.

Fixes https://github.com/dotnet/roslyn/issues/5343. I verified that this fixes the repro for both shared projects and cross-targeting dotnet core projects.