Capture better dumps when the contents of linked files don't match
In order to gather more information on
https://devdiv.visualstudio.com/DevDiv/_workitems?id=209299, we created
https://github.com/dotnet/roslyn/pull/16918 to capture the text of the
linked documents that aren't in sync. However, we tried to do so by
putting their text in local variables that were optimized away in dumps.

This change introduces a custom exception object that holds onto the
text of the conflicting documents and reports that exception.