Don't open an editable invisible editor when locating nodes
We implement GetStartPoint in a two step process: locate the point
directly from the tree, and then take that point and get an
EnvDTE.TextPoint from it. In the first step we were opening an editable
invisible editor for for no good reason, which was causing us to check
out files from source control. The second step does require an invisible
editor (since EnvDTE.TextPoint is implemented by the shims) but that we
were already opting out of an editable editor for.

Going through source control, we used to open an editor for the first
step because at the time we were grabbing tab settings which required
the editor to be open, but that's no longer the case.