Change assembly checksum caching
Part of building up our content file for a test description includes
calculating checksums for several thousand assemblies.  This takes up a
significant amount of time if done without optimization (~1 minute).  To
optimize this process RunTests caches checksums on a couple of levels.

One cache is based on file path.  The tool assumes the file system is
static for the duration of the run and hence caches checksums based on
file paths.

This cache was added back when the build outputed the majority of the
build to a single directory.  There are a lot of common assemblies
between our tests hence this was a very effective cache.  Since we've
moved to a more granular output structure this cache is becoming
extremely ineffective.

This change replaces it with an MVID based cache.  Calculating the
checksum of an assembly is relatively expensive but reading it's MVID is
cheap and it's an effective content key for PEs.

Local measurements show this saved us 55% on calculating the content for
the entirety of our test run set (about 20 seconds).