Ensure non-fatal exceptions are actually thrown
When a process crashes due to an unhandled exception or an explicit `Environment.FailFast()` or similar, the collected call stack and (possible) memory dump will be very close to the point of failure. This is obviously desirable when you need to diagnose an issue from crash reports.

When we catch an exception and report it as a non-fatal event, however, that collection may happen well after the failure. There are a couple of things we need to do to ensure we get the proper data.

First, we need to ensure that the exceptions we report have actually been thrown. The call stack within an exception is populated _when it is thrown_, not when it is created. As the analysis tools on the back end use the exception stack trace to bucketize failures, we currently have a large number of reports where we know something went wrong, but we don't know where.

Second, our reporting mechanisms are best called from within an exception filter. This is because the filter runs at the point where the exception was thrown, before control flow has passed to a catch block. If we collect a memory dump at this point it will preserve the stack frames that led to the exception which is very useful for debugging.

The resulting pattern is often a little odd: we throw an exception, report it through the exception filter, immediately catch it, and then do nothing in the catch block. However, this ensures we collect actionable data.