Fixes #18188
The bug arose because of an interaction of this line (new in VS2017)

``` c#
                _builder.EmitIntConstant(endConstant.Int32Value - startConstant.Int32Value);
```

in `SwitchIntegralJumpTableEmitter` and an incorrect implementation of `Int32Value` for constants of type `ushort` and `ubyte`, which is recorded as a separate bug in https://github.com/dotnet/roslyn/issues/18579. Specifically, the implementation causes improper sign extension of what should be treated as an unsigned value.

We never caught this before because we didn't rely on it. Constant folding in the compiler goes through another (redundant) code path that gets the padding vs sign extension correct. Because this is a serious regression in the compiler, I am hoping we can include this fix is some upcoming patch at our earliest opportunity. Therefore I have included a simple, local fix for the symptom, while leaving the larger issue of `ConstantValue` to be dealt with in a later larger change.