Set the rule set file when the command line changes
The Language Service needs to know the full path to the .ruleset file (if any) so it can watch it for changes and reload it as necessary. The legacy project system communicates the rule set path through the `IAnalyzerHost.SetRuleSetFile(string)` method. However, the new CPS-based project system doesn't know about this interface; it simple gathers the command line arguments and passes them to the Language Service to process accordingly. This change updates the Language Service support for the CPS-based project system to pull the rule set file path out of the parsed arguments and call `SetRuleSetFile` itself.