Ensure proper handling of a line-break following an incomplete string interpolation. (#18619)
Ensure proper handling of a line-break following an incomplete string interpolation.

Fixes https://devdiv.visualstudio.com/DevDiv/_workitems?id=405887.
The error recovery wasn’t including the line-break into interpolated string, yet wasn’t creating proper statement terminator token for it. The change ensures that the line-break is treated as part of the string for the purpose of the error recovery, this is what would happen if the interpolation was properly terminated with ‘}’ before the end of the line.