Make sure non-fatal exceptions have been thrown
We have the `FatalError.ReportWithoutCrash` method for reporting
non-fatal exceptions. There have been cases where this method is passed
a new, unthrown exception. These exceptions won't have a stack trace,
which makes the resulting telemetry and Watson reports more or less
useless.

Here we check if the exception has a trace, and if it doesn't we simply
throw and catch it before reporting it. This will generally be good
enough, on the assumption that the underlying error occurred (and thus
the exception created) near the call to `ReportWithoutCrash`.