Use a custom equality comparer when comparing metadata references
It seems like the reference set for CoreCLR tests is sometimes different, which
suggests that there may be a race in retrieving the set of standard references in
the static fields. This change should be resilient to differing object identity, so
it would work even if there is an race in setting the static fields.