Fix the double write of xunit.abstractions.dll
In order to leverage LUT testing we need to have both of the
follownig packages in our build.

- xunit.runner.desktop
- xunit

Both of these include xunit.abstractions.dll in different forms: content
and reference respectively.  This creates a double write violation in our
build.

I've spoken with the xunit team and at this point they believe the
feature is executing as designed.  The content use in xunit.runner.desktop
is meant to support 1.X clients who can't reference the package.

I'm still working with them to find a better way to exclude this but
for now this will remove the race in our build.

Issue #18573 will remain open to track the handling here and find
a better solution.