Create a NETCore-targeted Csc package
Creates a new package, Microsoft.NETCore.Csc, which contains a Csc
binary compiled against the netcoreapp1.1 shared framework. This can be
installed as a NuGet package and built using the dotnet SDK, as long as
the netcoreapp1.1 framework is installed.

This change also downgrades a few packages, like
System.Diagnostics.FileVersionInfo, that were higher than the
version available in netcoreapp1.1.

# Conflicts:
#	src/Compilers/VisualBasic/VbcCore/project.json
#	src/Tools/Source/CompilerGeneratorTools/Source/BoundTreeGenerator/project.json
#	src/Tools/Source/CompilerGeneratorTools/Source/CSharpSyntaxGenerator/project.json