Simplify how we go from a text-snapshot to a document in lightbulbs.
After investigating with jason, we've concluded this strange code was written
back in the days where roslyn workspaces and VS documents were not something
you could trust to be in sync.  So we had to have a lot of painful code to try
to detect and bail out when that happened.  This is definitely not necessary
now, and we can just use the simple and correct function we already hve for this
task.