Fix Release builds
The content produced by NuGet restore is independent of build
configuration.  Hence it needs to be stored in a folder that is
accessible to both Debug and Release.

This didn't work with the present structure of our Obj directory
because it put project folders under configuration directories.  That
mean there is no common parent for a project across all configurations.
To fix this we need to change our Obj structure:

- Old: `Obj\<Configuration>\<Project Name>`
- New: `Obj\<Project Name>\<Configuration>`