Fix BuildBoss
This changes our package verification logic to run inside of BuildBoss
instead of RepoUtil.  The verification is really easy now that package
references are 100% MSBuild based.  No reason to do this in RepoUtil
anymore.

As a result this also fixes most of the correctness leg.  Still need to
fix up the generators as they aren't deploying as EXEs anymore.  Need to
understand how to fix that.