Fix script for generating compiler code
This updates our script for generating and verifying the compiler syntax trees,
bound trees and error facts.

The complication here is we needed to move to fully runnable .NET Core
applications.  There was a lot of hacks / bad data laying around from
our previous hacky .NET Core app build story.  Now that we are on SDK we
can remove that and just do the right thing.