Fix crash in lightbulbs in projection scenarios.
A previous change of mine moved it so that we try to grab the user's selectoin as the first
thing we do when queried in lightbulbs.  This was to make it so we would not call back into
the UI thread from a BG thread.  Unfortunately, the code i moved was not safe to call if the
query was a projection buffer.  This wasn't a problem previously because there was a prior
check that ran that would cause us to bail out.

I've now reordered things so that the check comes first, and then we get the selection afterwards.