Fix Tests
The trait NetCore was being unconditionally passed as the trait option to our unit tests.  That mean all tests were being filtered
to those which had the trait NetCore which is a very small number.  Fixiyng it to only be passed when running the net core suites

Break was inserted via bad copy / paste in the following

https://github.com/dotnet/roslyn/commit/970cf1762ed422df17a7a22646be0d178d1c139f