Remove dependency on 3.5 framework
The Net20 framework projects were accidentally depending on 3.5
framework in order to build.  This is due to how they generate
resources.

By default when generating resources for a 2.0 framework project the
MSBuild task from 3.5 was being used.  This is not installed by default
with Visual Studio and only through the 3.5 framework.

This went unnoticed when moving to the new SDK because all of the
Jenkins machines and @jaredpar machines have it installed.

Setting the `$(GenerateResourceMSBuildRuntime)` property to
CurrentRuntime will use the GenerateResource task from the current
version of MSBuild.  This is how resources were being generated before
the migration.