Attempt to diagnose devdiv #384109
Sometimes completion filtering crashes in MatchesFilterText. From
inspecting the code, it appears that the only possible cause is that a
null completionHelper is passed in. The method that retrieves a
CompletionHelper sometimes returns null. This change will cause a crash
instead, allowing us to investigate why the Document was null, or try to
further determine why MatchesFilterText might crash.