Restore ability to deploy VSIX as a part of build
The VS SDK team is interested in tracking down the deployment errors
that we are seeing when deploying as a part of build.  This change adds
an option to make it easy to excercise that code path in builds.

Going to use this option to quickly toggle to the build based deployment
in order to mass test their fixes.