Introduce ReferencePathWithRefAssemblies parameter to CoreCompile target, and shim (#19146)
* Use refasm-only items for compilation

Use the new item provided by common targets to avoid recompiling a
project when a reference has changed only in implementation, not
interface.

Include a shim so that these targets can continue to work on older
MSBuild distributions. That keeps the NuGet-package-for-downlevel
scenario working, as well as making an easy transition period to the new
logic.

Requires Microsoft/msbuild#2039 to enable end-to-end use of reference
assemblies, but is compatible with earlier MSBuilds (without using
reference assemblies).

* Use CopyRefAssembly from NuGet package

The Microsoft.Net.Compilers package delivers UsingTasks to override the
"standard" ones provided by MSBuild's Microsoft.Common.tasks. That
should include the new CopyRefAssembly.