Use file path to hook up analyzers nodes in Solution Explorer
Fixes https://github.com/dotnet/project-system/issues/2062.

Currently we match a hierarchy node representing an analyzer in Solution Explorer up to the proper Roslyn `AnalyzerReference` based on the "display name"--basically, the name of the assembly without the extension. This mostly works, but you can run into problems if a project uses two different versions of the same analyzer, each with a different set of diagnostics. In this case expanding one of the analyzer nodes in Solution Explorer will show you the wrong set of diagnostics for one or the other, since they both have the same display name.

To uniquely identify an analyzer assembly we should use its full path rather than the display name. However, the project system initially didn't expose the full path through the hierarchy node. This has been fixed, and now the full path is one of the components of the node's "canonical name", along with other bits of information such as the project directory and target framework.

This commit updates our code to extract and use the full path instead of the display name.