Force load project were we don't have command line args in DPL
In the case where we AnyCode is unable to determine the command line
arguments for a project (likely because something is different about the
way design time builds are set up), fall back to forcing the project to
load normally so that users still get a good experience from IDE
features.

Fixes #19407.