Use ProcDump for dumps
Decided to go with procdump for getting our dumps.  While it's cleaner
and easier to create the dumps with a simple PInvoke call it's not a
complete solution.  The MiniDumpWriteDump API will always write the dump
in the architecture of the process calling the API.  Hence it would
create them for the architecture of RunTests, not the architecture of
the process being dumped.

In order to make it work I was going to have to buil two versions of
RunTests, or a simpler dumping program.  At that point it's just easier
to use ProcDump.  Even though I rather dislike an explicit download step
at the start of our CI runs.  Would be much nicer if sysinternal was
offered as NuGet package.

But the benefit here is clear so going forward with this.