Explicitly specify our Git dir to GitLink
When the baseDir is not specified GitLink attempts to probe for the .git directory.
This probing is done by looking at the first source file it finds in the PDB and
walking upwards.  This is a bit of a flawed strategy because generated files in
our Obj directory get referenced in the PDB but don't necessarily need to be in the
cone of the .git directory.

Fixing this by just being explicit about our Git directory.

https://github.com/GitTools/GitLink/issues/155