Clean up our post build step
This change has two effects to our build process:

1. Ports our GitLink fix from master
1. Creates an Index directory as a part of our post build work

Our index step today is very inefficient because it indexes virtually
every binary that Roslyn produces.  In actuality it should just be
indexing the ones that we ship.  The difference is literally gigs of
data.

The index diretory solves this by creating a simple layout which
contains only the data that actually needs to be indexed.  The index
step can simply target every file in this directory.