Exclude projects for aspx/cshtml/vbhtml files
Fixes [VSO bug 441171](https://devdiv.visualstudio.com/DevDiv/_workitems?id=441171).

When mapping an `IVsHierarchy` for a project node in SOlution Explorer
to a Roslyn `AbstractProject` we try to match the properties on the
node's hierarchy against properties on each project's hierarchy. In the
case of a Web Application Project we may still end up with multiple
matching Roslyn projects: one for the "main" project, and one for each
open aspx/cshtml/vbhtml file. We're only interested in the main
project, so this change excludes any projects with a
`ContainedDocument`. This may involve checking a lot of documents, so
we only do it if we must.