Fix bug in closure rewriting
When lowering local functions that capture struct frames, the frames are
added as ref parameters to the lowered method. Normally, when the local
function captures variables from multiple levels up the intermediate
closures are also rewritten to take the ref parameters. However, if one
of the intermediate closures captures no variables, the closure
conversion code can skip adding ref environments to that closure. This
can cause the inner-most closure to no longer have access to all of the
ref structs it expects, which crashes the compiler.

Fixes #18814, #18918