Log telemetry if we see ITextBufferEdit.Apply() throwing exceptions
We believe that the primary cause of this exception (some issues around
projection buffers) was worked around in d8f8dad3dd58 but we're still
seeing issues so that's not the case. I'm also further not convinced
that change helped at all, because in many case we already fork
buffers to work around the same issue.