Fix crash in VB type inferrer
One path in the VB type inferrer assumed that invocations of Select() would always have an argument list, leading to crashes. We now check that it is not null before accessing it. This is the only unguarded use of this property in the type inferrer.
Fixes  https://devdiv.visualstudio.com/DevDiv/_workitems?id=431509&_a=edit&triage=true