Crash if debugger contenttype does not match
We've received watsons that indicate that a VB file is being set as the
"context buffer" for C# debugger intellisense. It appears that this
would only happen if the debuggered called us with a buffer of the wrong
content type. Add an assertion that the content type is correct.

Related to
https://devdiv.visualstudio.com/DevDiv/_workitems?id=397711&_a=edit&triage=true