Include Git commit in our NuGet packages
This change adds the git commit which produced the NuGet package to the
description section. This makes it easy to both

- Determine what source is included in a NuGet
- Determine if a given change is included in our NuGet

Lacking this only way to understand which commit produced a NuGet is to
download the NuGet, crack open the binary and look at the version we
embed into the PE. This allows us to see the version on the NuGet page
itself.