Don't subscribe to file watching for analyzers until we have to
Before this change we were immediately trying to add file watchers
on the UI thread when we added a new analyzer. This can get delayed
because of slow disks, or because the file watcher service's lock is
held by a background thread. There's no reason to do it until we are
actually observing and loading the analyzer, which typically happens
on a background thread.