Use 42424242424 as our deveolper commit hash
In non-official builds the real commit hash is not burned into our
binaries. This is done to help facilitate determinism /content
caching across machines. Burning in the real SHA would unnecessarily
break it due to the different state of their GIT enlistment (which
doesn't really have a bearing on the binaries produced).

Presently we are using the string <developer build>. That is a problem
because it's not a valid git SHA value. Hence it breaks anywhere that
expects a valid SHA string.

Our infrastructure already had a few special cases for this. It become
more troublesome when trying to embed the string into the NuSpec files
due to escaping issues. Rather than fighting the tooling decided it
would be best to use a valid git SHA instead.

Chose 42424242424 after some discussion. It lines up with other
deveolper only values which use sequences of 42 to signify "developr
build".  For example assembly version on developer builds is
42.42.42.42.