Disable implicit transitive references in VSIX projects
By default the new SDK will cause projecs to reference the full
transitive closure of project references, not just the direct
references. This is fine in most cases but caused breaks in our VSIX
projects.

Our VSIX projects often reference project with specific extra metadata
to control whether or not their contents are included in the final VSIX.
That extra metadata is not applied to projects which are implicitly
referenced by the new SDK. That caused our VSIX to bloat in size and
have functionally incorrect contents.

Fixed this by disabling implicit transitive references for our VSIX
projects.

https://github.com/dotnet/sdk/issues/1366