Fix our NuGetPackageToIncludeInVsix support
The NuGetPackageToIncludeInVsix element is used to make the NuGet assets
we use in a VSIX very explicit. It depended on the nuget items having
the metadata element NuGetPackageId. The 1.X New SDK no longer has this
piece of metadata and hence we had to pick a new approach.

After some discussion decided to simply look for the assets which are
coming out of the NuGet packages folders that we expect. It's
essentially the same check.

Needed to encode this in an MSBuild task though as the MSBuild
transforms can't handle that level of complexity.