Report as being on the foreground thread in more cases
Prior to this change, we would only report a call into
ForegroundThreadAffinitizedObject as being on the foreground thread if
the calling thread was the same thread as the one believed to be the
foreground thread at the time the object was instantiated. But,
sometimes our understanding of the foreground thread changes (because
xUnit changes the foreground thread during test runs or because
ForegroundThreadAffinitizedObject gets statically constructed from a
background thread and we later determine that there's a more correct
"foreground thread"), and we now allow calling threads to match the
thread currently understood to be the foreground thread.

This change also records the StackTrace of every ForegroundThreadData
constructed for debugging purposes (so we can see when the
ForegroundThreadAffinitizedObject is statically constructed from a
background thread).