Don't fetch every PR ref for every PR job
Newer versions of git (I've seen this behavior on 2.13.1.windows.1 and
have not seen it on 2.11.1.windows.1) are much slower at writing out local
copies of refs after a fetch. Given that we fetch about 10K refs when
running each PR job this can lead to a massive slowdown in fetch times (on
the roslyn integration job, which has a newer version of git on the
workers compared to the other machines, we see fetch taking ~20 mins).

This changes all the Roslyn jobs to fetch just the PR they are going to
test, which speeds up the fetch significantly.