Add Scope tree data structure for closure conversion
This commit adds a tree-based data structure for analyzing capture and
lowering scopes for closure conversion in C#. The tree node is called a
Scope and it contains a set of nested Scopes, along with information
from the bound tree about what variables are introduced and captured in
each scope.

This commit adds only the initial pieces of the scope tree, designed to
replicate the functionality of the ScopeParent, VariableScope, and
CapturedVariablesByLambda dictionaries in the current closure conversion
code. The current implementation is also meant to be completely
equivalent, so there are asserts in ComputeLambdaScopesAndFrameCaptures
that verify that the resulting LambdaScopes, NeedsParentFrame, and
ScopesThatCantBeStructs data structures are identical after both forms
of analysis.