Use source linking in our PDB files
This makes a couple of changes to how PDB source server information
works in our repo:

- Uses source link vs. GitLink
- Enables for all built DLLs vs. just shipping ones.
only
- Enables for virtually all builds vs. just official builds.

The main item to discuss is the difference between source link and
gitlink:

- Gitlink is a post processing step on our PDB which replaces document
paths with raw URLs. This is understood by both VS and windbg.
- Source link inserts a JSON file to the PDB during normal compilation
which describes how to map files in the PDB to Github URLs. This is
understood by VS only (windbg support pending). This is a format
specifically supported by our compilation pipeline.