Remove unused fields
When examining types in a file for a `DesignerAttribute` we record whether or not we encountered an error, and whether or not the attribute is even "applicable"--basically, whether or not we can even find the `DesignerAttribute` type to begin with. However, neither of these values is actually used anywhere, so just remove them.