Fix a build warning in SyntaxVisualizerExtension
Need to use the correct version of MS.VS.ComponentModelHost to line up
with the rest of our source code.

Additionally now that this is building as a part of our standard build
the warning can't silently come back. It's build under the /warnaserror
flag and hence will trigger a build break in the future.

closes #17325