Use multi-targeting to control compilation helpers
This changes our core Roslyn.Test.Utilities project to use
multi-targeting to control how we create our core compilation helper.
Previously this was done by having a separate runtime specific assembly
which was probed at runtime.

The intent of this change is to make our build more correct here. I can
now target just Roslyn.Test.Utilties with a net46 framework and get a
runnable asset out of it. This is necessary in order for our unit tests
to successfully run when moved to multi-target.

Presently this makes it multi-target across net46, netcoreapp2.0 and
netstandard1.3. The impl in netstandard 1.3 simply throw for this
specific API. Over the long term it make make sense to pull this out to
a separate library that only targets net46 + netcoreapp2.0. That
requires a much deeper investment in multi-targeting though. Decided to
start smaller for now.