Slightly more efficient way to emit a reference to default expression
Basically addressing a TODO in codegen.
To pass default by reference we need a temp and to make a default value we need a temp (often mapped to same slot)

We could just pass the reference to the second one. No need to copy between temps.
Besides, they are often mapped to the same IL slot resulting in clearly redundant load/store with the same local.

The scenario gets more common with "ref readonly" so here is the fix.