allowNewCompilerDiagnostics description clarified
bool allowNewCompilerDiagnostics param has an ambiguous name and description:
/// <param name="allowNewCompilerDiagnostics">A bool controlling whether or not the test will fail if the CodeFix introduces other warnings after being applied</param> 
In this case 'allow' should be interpreted as 'the assertion will allow (ignore) new warnings', but may also be interpreted as 'the verifier will allow them to be added to diagnostics list'. This inverts the meaning of boolean parameter. 
New description makes it clear, what the parameter means:  
/// <param name="allowNewCompilerDiagnostics">A bool controlling whether or not the test will ignore any new warnings intoduced by CodeFix application; 'true' means ignore any new warning during assertion; 'false' means the test will fail if new compiler warnings are introduced</param>