Add fields as members of synthesized structs
In order for EnC and other mechanisms to work we have to add synthesized
members to a list in the CommonPEModuleBuilder for a compilation (these
synthesized members are then queried as part of compilation stages).
If those members are struct fields, we don't add them to the struct
definition, only to the list of synthesized members. This works for
emit, since we explicitly emit everything in the synthesized list, but
it doesn't work for any compiler pass that examines the members of the
struct for semantically meaningful reasons.

This is the case for the CaptureWalker for async and iterator
expressions. The walker checks the members of structs when a field of a
struct is assigned to see if the struct has been assigned piecewise
(each of its members has been assigned individually). If so, it will
mark the entire struct as assigned. By not including synthesized fields
as proper members of the struct type, the assignment pass believes that
many fields have been assigned that have not, and thus marks the full
struct as assigned, leading to losing track of variables captured across
await/yield statements.

This PR fixes the problem by adding the fields to the
SynthesizedContainer, but exlcuding them from emit, since we should
continue to use the emit mechanism used for CommonPEModuleBuilder.

Fixes #21409