Address feedback from the design meeting and only expose the top level ITranslatedQueryExpression IOperation for v1.
We longer expose any special IOperation nodes for query clauses/operators, but just the underlying query plan. For v2, we hope to enhance the APIs to expose a richer query API.

Another design decision was to rewrite VB query lambdas into regular lambda, so we have parity between C# and VB operation trees.
This change refactors the VB query lambda local rewriting code to be consumed from the VB operation factory for the same.
The rewriter introduces a large number of DAGs, and we handle them in the operation factory rewriter by cloning the duplicate bound nodes in the lowered tree.