Move InlineRename to EditorFeatures.Wpf
Although this could be split out more precisely, for now we'll just
move the entire feature to the WPF layer. IEditorInlineRenameService.cs
is left behind however, as that interface is implemented by TypeScript
and other languages. For now I'm considering EditorFeatures.Wpf to be
an implementation assembly for now and so interfaces implemented
should still live in the base assembly.

In CSharpEditorFeatures and BasicEditorFeatures, there was the
implementation of IEditorInlineRenameService which had no language
specific details other than the language name the service was exported
for. I have moved those into EditorFeatures.Wpf rather than have
additional assemblies for language specific stuff, since I anticipate
that to be nearly non-existent.