Remove our custom checking for high contrast mode
The editor provides an API if it's in high contrast mode which can
happen either because the system is in high contrast mode or they've
chosen a Visual Studio theme that is higher contrast. This eliminates
an icky WPF dependency.

We stil disable generation of the classification if we are in this
mode, because editor ignores the tags we'd produce anyways when in
contrast mode.