Fix build parallelism issue with documentation file generation
Use GenerateDocumentationFile=true instead of DocumentationFile=
$(IntermediateOutputPath)$(AssemblyName).xml. This lets the SDK
pick the appropriate path.

At the point where the explicit DocumentationFile was being evaluated,
the intermediate output path had not yet gotten the TargetFramework
appended. As such, parallel builds of the different TargetFrameworks
of a multi-targeted project were trying to write the documentation xml
to the same intermediate location.