Move QuickInfo to EditorFeatures.Wpf and EventHookup to CSharpVisualStudio
This is a really ugly refactoring that removes the WPF APIs from
IDeferredQuickInfoContent and adds a service in EditorFeatures.Wpf that
converts from them to WPF elements. It's a straightforward refactoring
and everything still works, but it's not pretty at all.

EventHookup also worked with QuickInfo directly, and thus had to be
moved out of CSharpEditorFeatures. There isn't a CSharp-specific
EditorFeatures.Wpf assembly, so it's going into the Visual Studio layer
for now.