Remove ForegroundThreadAffinitizedObject's dependency on WPF
ForegroundThreadAffinitizedObject took a WPF dependency so it could
create a lower-priority DispatcherSynchronizationContext, which
is something we can set up during our package load where we were
already forcing the foreground thread information to be initialized.
We also have to strengthen an assertion since we can't check for
application initialization anymore.