Fix issue where tests were creating TestWorkspaces during test disposal
TestWorkspaceFixture is a type you can use to have a single
TestWorkspace be shared between multiple tests. To ensure that the
TestWorkspace doesn't end up with views being opened between tests,
many tests in their dispose method call CloseTextView(). This was
calling GetWorkspace(), which creates the workspace lazily, so in some
cases we'd create a test when we didn't expect it. Since this was
happening in xUnit's Dispose() handling, there wasn't a synchronization
context which is now required due to 0911290b06184eb.