Fix inconsistencies in the way reachability is handled in pattern switch (#22335)
* Improve consistency of handling constant null switch expression
Fixes #20210
* Rationalize reachability for pattern switch.
If a single case could handle all inputs, e.g. has the
same constant value, a wildcard pattern, etc, then
the default label is considered unreachable.
The concept of reachability, which takes the value of the
switch expression into account and can produce warnings,
is separate from the concept of subsumption which takes
the full set of patterns into account and can produce errors.
* Document breaking change in pattern switch.
Alse change formatting as suggested in code review.