Add an end-to-end perf test for Roslyn on CoreCLR
This is a first step in getting a perf suite running for the compiler.
The current suite only supports running an end-to-end csc.exe
compilation against an old copy of the CodeAnalysis.dll source code, but
it can be augmented in the future to support more runtimes, platforms,
and types of tests. The underlying perf utility is BenchmarkDotNet,
although this test uses a heavily mocked-out path that skips over a lot
of BenchmarkDotNet functionality in order to run against an external
process.