Move all build logic into build.ps1
After this move the `-buildAll` switch will now build the entirety of
the Roslyn source tree. Previously it didn't build a number of items
related to VS insertion. That layering was arbitrary and mostly due to
limitations in our signing process. Now it can all be done in a single
phase.