Upgrade to the Visual Studio editor 15.6.161-preview packages
This lets us remove many stated dependencies on
Microsoft.VisualStudio.Text.Internal and dependency on
Microsoft.VisualStudio.Platform.VSEditor.Interop. Both
is a stated dependency of Microsoft.VisualStudio.Platform.VSEditor
which means we don't get any benefit for trying to state it -- and it's
actually hard since it has a different version anyways.