Reuse proxy fields in async state machine
We currently provide every local with its own proxy, even when locals are in disjoint scopes.
That is done for the purpose of debugging, especially EnC.

We do not have to do this in Release ( /o+ ) and reduce the size of async state where possible through reusing of backing fields.

NOTE: reuse does impact debugging experience when reuse happens - locals that are mapped to recycled fields generally cannot be examined. Otherwise debugging experience stays the same.

Fixes:#15290