Upgrade CoreCLR toolset to work on Windows (#23110)
The old package had a bug in RunCsc.cmd that didn't properly
quote the dotnet path. When the SDK is upgraded to a sufficient
version by VS, the full dotnet path will be used, which contains
spaces and fails without proper quoting. This updates us to a package
which contains almost the same compilers, but includes the fix for
quoting the dotnet path properly (in PR #23034).