Enable transitive project references
When moving to the new SDK I initially disabled transitive project
references because it was causing issues building our VSIX projects. Had
some time this week to sit down and work through the problem which comes
down to the VSIX indirectly referencing a multi-targetted project. This
is a known bug in the SDK that can be worked around by doing the
following:

- Explicitly referencing the project
- Using AdditionalProperties to specify the TF to use
- Using Name to work around inability of VSSDK to infer Name

The SDK bug is here.

https://github.com/dotnet/sdk/issues/433