Restore toolset packages to Obj directory
The toolset packages were all restoring to the source directory. This
violates our build principal of builds should never write to the source
directory.

Changed to restore to the Obj directory as the rest of our code does.
Temporarily went back to the old MSBuild syntax. This will move to the
new MSbuild syntax with the rest of the repo.

The general principal here is that source directories should be
immutable. Ideally builds should be done with the write permissions set
to deny on the entire source tree. Having an immutable source tree helps
ensure that:

- Builds are repeatable: Any mutation, even simply new files, means it's
new input for the next build to see and potentially alter behavior.
- Builds can run in parallel: The source tree is read many times during
build. Any mutation to the tree can impact parallelization as order of
reads and writes must now be taken into account.