Fix QuickInfo for F#
The QuickInfoDisplayDeferredContent type was somewhat evil: it had a
set of properties that were intended as test-only accessors, that would
give you access to fields of type IDeferredQuickInfoContent but would
cast to the expected concrete type. In a previous refactoring I
started using those properties in the product which breaks F# because it may
create different implementations from IDeferredQuickInfoContent.

My fix here is just to delete the properties and clean up the test
helpers to do explicit casting. By refactoring the helpers a bit the
test complexity doesn't change and eliminates the trap in the first
place.