Use binary log in microbuild
This enables the generation of binary logs for our build when running in Microbuild.
Having the logs available helps in post-build debugging of errors. We tend to get a
number of microbuild only failures that are impossible to track down without logging
like this.

Once this is merged i will be adding a task to our VSTS build definition to publish
these post build.