Turn off loc for SyntaxVisualizerExtension.csproj
Related to https://github.com/dotnet/roslyn/issues/23566.

We've basically forced all projects to use the SDK approach to
localization, but this project doesn't use the SDK. As a result, it
doesn't produce satellite assemblies, and the build fails when we try to
copy the non-existent files to the output directory.

For the moment, just turn off XLIFF-based localization.