Suppress xlf on templates.csproj
The templates project was not being picked up by xlf previously due to
resx files being included in `<None>` items in the project file. Once we
moved to the new SDK here they were turned into resources by virtue of
the EnableDefaultEmbeddedResourceItems property. Disabling that for now
as it gets us back to the state before this change. Filed a bug to track
re-enabling it at a later time.

https://github.com/dotnet/roslyn/issues/23649