Change local function definite assignment
The LDM has decided that the following rules should be in effect when
calculating definite assignment for local functions:

  1. The entry point to a local function is always reachable.
  2. Variables captured in local functions are definitely assigned if
  they are definitely assigned in all branches into the local function.

It turns out these rules were almost exactly what the compiler already
implemented, but there was a bug in captured variable detection that
meant that variables captured in lambdas within local functions were
sometimes not counted as captured. This change fixes the bug around
capturing, which should cause the compiler to conform to this
specification.

Fixes #17829