Fix hang in compiler server
The dotnet CLI has been seeing a hang from VBCSCompiler in their build
on non-Windows platforms. The cause is this:

    1. dotnet build starts a child process to do `dotnet pack` and
    redirects its output to a buffer.
    2. `dotnet pack` starts a child process to do `compile`.
    3. `compile` starts vbcscompiler because one is not already running.
    4. VBCSCompiler inherits the output handles from (1).
    5. VBCSCompiler process finishes and processes (1) and (2) exit.
    6. The parent `dotnet build` process  is waiting for (2) to exit,
    which it has, but also to see EOF on its output stream. Because that
    stream has been captured and held open by VBCSCompiler, the EOF
    never comes.
    7. `dotnet build` hangs.

This change fixes the problem by creating new streams for input and
output for the compiler server process. This isn't quite what we do on
Windows -- on Windows we create invalid handles for the input and output
-- but it's as close as we can get using the portable APIs in
System.Diagnostics.Process.

Fixes #23734