Remove obsolete duplicate APIs from type symbols. (#23888)
This change removes the following APIs:
- TypeSymbol.BaseType
- TypeSymbol.Interfaces
- TypeSymbol.AllInterfaces
- NamedTypeSymbol.TypeArguments
- TypeParameterSymbol.ConstraintTypes

These APIs simply delegate the work to APIs with "NoUseSiteDiagnostics" suffix and are not supposed to be used inside compilers. They were left in the code-base because they used to be part of the public surface and the names were nicer, but they are no longer part of the public surface.