Remove IVisualStudioWorkspaceHost and related interfaces
This interface was originally introduced to support CodeLens, when
we needed the ability to create a Workspace that was out-of-proc but
connected to the in-process project system interfaces. This created
the abstraction of how we pushed updates to the workspace, and came
with the rule that the VisualStudioProjectTracker was unable to
ever touch a workspace directly.

Today, that's not how we build our out-of-proc components. We now
have an in-proc workspace, and we mirror that in-proc workspace
via a pull model to the out-of-process workspace. This meant that the
interface was nearly pointless, except it was being used for the
out-of-proc workspace setup to mirror some persistence settings over.
Providing a new interface to do that directly is simpler and also
much cheaper, as it doesn't come with the overhead of pushing
workspace changes to the remote host that were being ignored.

This has the direct perf benefit of removing some delays around setting
up the remote workspace since we're not computing useless data. It
has longer-term benefits to aid in further refactorings, as we now have
workspace updates going directly to the in-proc workspace, which
we can start to do in a free-threaded manner.