Use NET46 for CreateStandardCompilation
This changes the CreateStandardCompilation helper to use the net46 version of
mscorlib. Previously it was using the net40 version.

This method is a helper to create a minimal compilation across both our desktop
and coreclr test executions. The coreclr version uses netstandard20 and hence
using net40 created a substantial API gap. This better eliminates the gap.

Note: I realize net461 is a slightly more correct choice. However that ref
assembly isn't checked into our code yet and I wanted to avoid a giant resource
update here unless we need it. Our tests focus on the intersection of the
two API sets hence it's unlikely to be an issueo here.