Add a script to check the Loc status
Add a script to check for unlocalized resources. It checks all .xlf
files in the src directory for `state="new"` or
`state="needs-review-translation"`. These attribute values appear on
elements representing resources that are new and have not yet been
translated, and resources that have previously been translated but are
now out-of-date with respect to the original English. If we find one of
these strings we print the file path and, later, return 1 to indicate
the check failed. Otherwise we return 0 to indicate that no untranslated
resources were found.

A proper examination of the XML would technically be more correct, but
this is simpler and more than adequate.