Add a Jenkins job to check loc status
Add a Jenkins job to check the loc status of a branch. If the branch
contains no unlocalized resources, the job will pass. Otherwise, it will
fail. This allows us to use Jenkins as a dashboard to see which branches
are in need of localization.

By default we don't want to run this on PRs as those will often include
new, unlocalized resources and they should never be blocked on that.
However, we want to leave open the option to enable it for PRs once a
branch has entered escrow; that way we can avoid taking changes with
unlocalized resources (or at least we'll know they need to be
localized).