Make a perf tweak in the DataTipInfoGetter
Often, these only require fetching the syntax tree, which we now do
synchronously. Since this is called on the UI thread, it's good practice
not to call GetSyntaxRootAsync, as if we do have to recompute it it
means we're having the UI thread wait for the thread pool which
might be backed up. By having the UI thread do it directly, it'll always
be faster.

I noticed this while working with @ivanbasov investigate some other
perf issues. Awhile back we did an "async everything" approach, and
have since learned that was perhaps overly aggressive, at least when
that's stuff that's on the UI thread. We saw a user complaint of this
taking longer than it should have, and this might help a lot, and won't
hurt at all.