Remove ExtensionInstallation* properties
These properties are used at build time to generate a catalog file that
tells the setup engine to install the extension to a specified location.
However, this extension is a per-user extension so the catalog file is
ignored. Additionally, it appears this extension is only used during F5
anyway, so the VSSDK would ignore the properties in the catalog there
too.

The reason this fix is needed is that in future versions of the VSSDK,
building a per-user extension with an ExtensionInstallationRoot of
CommonExtensions will cause an error (in order to reduce confusion at
install time and to protect ourselves in case the SetupEngine begins to
install per-user extensions).