Manually load selected MSBuild dependencies from runtimes/ directory
Certain dependent assemblies (System.IO.Pipes,
System.IO.Pipes.AccessControl) are operating system dependent and thus
end up in the runtimes subdirectory next to the build task.
Unfortunately, there is no mechanism for automatic assembly resolution
of plugin DLLs like MSBuild tasks, so we have to implement something
similar ourselves. This change hooks assembly resolution and, when we
see a well-known DLL request, we redirect it to the appropriate path.