Fix compiler server on CoreCLR and enable bootstrapping on CoreCLR build (#24239)
This PR fixes the CoreCLR compiler server on Windows and changes our bootstrap 
build to act as a test case. The build was previously failing during pipe ownership verification
because the required API was not available through reflection. This change
uses the System.IO.Pipes.AccessControl API to do pipe ownership verification
instead. However, System.IO.Pipes.AccessControl is operating system dependent 
and thus ends up in the runtimes subdirectory next to the build task.
Unfortunately, there is no mechanism for automatic assembly resolution
of plugin DLLs like MSBuild tasks, so we have to implement something
similar ourselves. This change hooks assembly resolution and, when we
see a well-known DLL request, we redirect it to the appropriate path.

Fixes #20900 and #24072