Include satellite assemblies when publishing for .NET Core
There was an issue in the custom PublishWithoutBuilding target that aims to
publish the project without rebuilding anything. It has the same dependencies
as the built-in Publish, but replaces build with Just ResolveReferences having
BuildProjectReferences=false. However, this caused BuildOnlySettings (a
dependency of Build) to not run, and therefore BuildingProject was set to false.
That put ResolveAssemblyReferences into design-time mode where FindSatellites
became false.

This fix is required so that the .deps.json that is generated contains satellite
assembly information. Without that, satellites will not load load on  .NET Core.

There was also a missing dependency on PrepareForPublish, which didn't cause any
issue (yet), but it's now added for completeness.

Also removed a nearby workaround for an unrelated v1 SDK issue with satellite
assemblies. It is no longer needed as Roslyn requires v2+ SDK.