Don't compute project data for the remote workspace's WorkspaceHost
When we are initializing the remote workspace, we connect an
IVisualStudioWorkspaceHost as a way to get notifications about the
solution being added and information being changed. The implementation
of this for most methods is a no-op, but we were still spending a lot
of time computing data to push to those methods that then ignored it.
This PR adds an opt-out flag to save some processing time.