Notify the remote workspace of working folder paths in the background
This pulls a .Wait() call off the UI thread in solution load scenarios,
since it's not urgent that we notify the OOP process about the working
folder path. We still ensure the unregister/register ordering is
maintained by ordering the Tasks, and if we have to synchronously shut
down the remote persistent storage, we will still wait. That is a rarer
gesture (it happens on solution rename) and is something we can address
if that's needed.