Add cast to prevent the dynamic binder from doing the cast
IVsRunningDocumentTable4.GetDocumentData() returns dynamic, and we
were casting it to an interface. This was going through the C#
runtime binder to do so. We don't need that, and an unrelated change in
VS meant that Microsoft.CSharp.dll isn't ngened, so this was far more
expensive than it should be.

@olegtk and @sharwell get credit for the fix.

Co-authored-by: Oleg Tkachenko <olegtk@microsoft.com>
Co-authored-by: Sam Harwell <sam.harwell@microsoft.com>