Rename to CompileStandardAndVerify
This unifies our API naming convention. When standard is in the name
then we are using the netstandard2.0 API set and when it is not then it
is the simple desktop mscorlib46.

Both variants are necessary for CompileAndVerify style functions because
many of our tests that use it can only run on desktop or simply need to
have some of the netstandard2.0 API set missing.