Use LHS of ref-assignment for ref-readonly
The spec for ref-reassignment is that the LHS of the ref-reassignment
expression is the ref-readonlyness of the LValue produced by the
expression. That is, if the LHS of a ref assignment is readonly, the
resulting LValue is as well.

This PR also fixes a related issue for ref-like assignments, where the
resulting lifetime of the assignment was decided based on the RHS,
instead of the LHS.

Fixes #24874