Fix SyntacticClassificationTagger's handling of workspace changes
The SyntacticClassificationTagger builds a queue of work items. When
it sees a workspace change of ProjectChanged, it sees if it needs to
enqueue a new parse due to optiosn changing. If it sees a DocumentChange
it also enqueues a new parse if it's the document (note: in active
context) we are tracking. In all cases (and after these two) it also
enqueues a task to ensure the _lastParsedDocument field is brought
forward to the current version. That had asserts (and assumptions)
that in that case the contents of the text editor would never change.

That assumption is not correct for two reasons:

1. If a linked file is changing, we raise two DocumentChanged; both
   present the same OldSolution and NewSolution. If the non-active
   context gets its event raised first, then we would try to migrate
   the _lastParsedDocument forward, but that also is bringing forward
   a content change too.
2. If a SolutionChanged event was raised (say for a refactoring) that
   also changed a document, we could get a document changing as well.

The code that migrated _lastParsedDocument, since it presumed the
content didn't change, would then only set _lastParsedDocument but
wouldn't set _lastParsedSnapshot. This meant things were out of sync,
at least until the next parse that would reset them. I suspect this
might be the cause of #23074.

As a fairly tactical fix, if the snapshots are different, I don't set
_lastParsedDocument but rather enqueue a new parse operation. That
way the two _lastParsed fields don't get out of sync and we still
will have a parse going that wil (eventually) refresh everything.