Enable indexing of `fixed` fields without pinning.
Turns out the reason why we need pinning is the artefact of our code generation strategy - it is actually unnecessary to convert to unmanaged pointer before indexing.
Same unsafe pointer math is equally applicable when we have the receiver in the form of managed pointer. However, when it is managed (GC-tracked), we do not need to pin it.

Therefore, when a `fixed` field is used just for indexing and the unmanaged pointer is never realized, the requirement for `fixed` statement is unnecessary, regardless whether the receiver is movable or not.

NOTE:
- it is still possible to pin `fixed` buffer. That would be useful if one needs to do some elaborate pointer math or pass the pointer to the native API.
- It was always allowed to index in unmovable contexts and it was always unsafe, and still is  - in particular no range checks are performed.
- the only restriction that is gone is that you can use indexing if receiver is movable as well.
The operation is just as fast and just as unsafe as in unmovable context - in particular it performs no range checks.