Moved C# command line tests to net461
The C# compiler is currently required to target net46. This meant the
command line tests needed to target net46 as well as the compiler was
copied into the same directory as the tests. This meant they shared
facades and hence had to have the same target framework (else the
compiler would fail with binding redirects).

This changes the compiler to be copied to a sub-directory below that of
the command line tests. Hence they no longer share facades and can
target different frameworks.