Remove NFW causing delays
Remove a non-fatal Watson report that was triggering frequently and
causing UI delays. We expected that all references to a project would be
removed before the project itself was unloaded, and this report was
added to help us diagnose the rare occasion when that did not occur. It
turns out that it happens quite frequently: when a solution closes, when
a project changes externally and needs to be reloaded, and when a
project is deleted from the solution. We're getting a lot more of these
events than expected, and worse, collecting the Watson data sometimes
causes long pauses in the UI.

We do not have a concrete plan for dealing with the underlying issue,
but the data we've already collected from these reports demonstrates the
scope of the problem and collecting more won't help at this point. Since
the collection itself is causing problems, it's time to just remove it.