Remove abstract methods in AbstractProjectCodeModel
Rather than using inheritence, we only need one factory method. By
favoring composition here this will set us up to remove a bunch
of derived types.

This commit is admittedly strange, as in some cases we now have a
child type being created in derived types of AbstractProjectCodeModel.
This is just to move things around, and it'll get better in the next
commit.