Introduce an interface IProjectCodeModel for ProjectCodeModel
This interface is able to live in our ServicesVisualStudio project
and that's what the project systems are able to interact with. This
makes the API more like an "optional service" that they can create.
Down the road, we can just have an IProjectCodeModelFactory that
keeps the implementation entirely separated out.