Move to the 2.1 SDK
This moves Roslyn to use the 2.1 SDK.

This may appear to be moving the SDK version backwards from 2.2 to 2.1.
In reality though it is a move forward. There was a brief time when the
SDK switched their versions to 2.2 before reverting back to 2.1. Roslyn
took a dependency on SDK in that small window.

More details available here:

https://github.com/dotnet/announcements/issues/57