Rid ourselves of IProjectCodeModelProvider, and collapse some method hierarchies
Our CodeModel implementation would go to AbstractProjects and then
see if the implemented a special interface IProjectCodeModelProvider
to fetch their ProjectCodeModel. This abstraction was pretty leaky:
since we're already assuming carnal knowledge of AbstractProject,
we can just make it a property that might be null and be done with it.
This means we can push some code that is in very derived types (and
duplicated) down into a single shared type.

At this point, all project system calls flow through IProjectCodeModel,
and the only tangling back to AbstractProject is just to do the lookup
from project -> IProjectCodeModel. We can now introduce an interface
IProjectCodeModelFactory that just maintains that mapping and then
suddenly things are really clean.