Make unit tests pass again
This fails because in unit tests (unlike before) we are now creating
a ProjectCodeModel. And now we're calling OnSourceFileRemoved when
we try removing files, but the logic to create an actual
CodeModelProjectCache is going to fail. If we don't have a cache, that's
fine for these because we're only trying to update documents we've
already handed out. If we don't have a cache, there's no documents to
fix.