Remove Portable server (#25449)
This code was built as a prototype for porting the compiler server to
other platforms. It has now been ported to CoreCLR and multiple OSes
through multitargeting. The portable server cannot meet our security
requirements and will never ship, so it's continued existence in our
code is merely confusing to anyone who happens upon it.