Key container signing fix (#25426)
This change makes the legacy behavior contingent on the runtime of the
compiler. When the compiler is running on the desktop runtime, the
legacy signing code will be used. When running on CoreCLR, the new
portable signing code will be used. This fixes key container signing for
almost all scenarios (signing using a key container when the compiler is
running on the desktop framework). Signing using a key container on
Windows using a compiler running on CoreCLR is still broken, but that is
a far rarer case.

This change also fixes delaysign, which would full sign the file if the
keyfile passed to the compiler contains a private key.

Fixes #25340
Fixes #25424