[SQLite] Fix continuous probing for sqlite
The static constructor of SQLitePersistentStorage would throw an
exception, which would then get logged by the exception logger.

This changes the following logic:

* The database used to be deleted if e_sqlite was not found.
* The sqlite persistent storage is not used if e_sqlite is not found.
* The database ownership lock is released if any exceptions happen
while trying to instantiate the storage.

Fixes #24042

Extra story:
There is a architectures of e_sqlite not existing for some platforms,
i.e. ARM64. Therefore, in MonoDevelop, it would throw a
TypeInitializationException on trying to initialize the SQLitePersistentStorage.

In this case, what would happen is:
* The persistent storage service would create the lock file.
* It would throw in the static constructor, without having passed
ownership
* Thus, the lock file is created by the service, not held by anyone
specifically
* The exception wasn't accounted for, so the storage file would be
deleted
* We now have a lock file present and no storage created
* Roslyn will keep trying to create the persistent storage on every
subsequent try, ending up in a performance issue caused by continuous
IO on trying to get ownership on the lock file, but it can't take it
because nobody released the old lock

Therefore, shortcircuiting and disabling persistent storage in case the
lib is not found is better.