Move all build diagnostic info into Logs dir
This moves all of our build diagnistic output into the same folder:
Binaries\Configuration\Logs. Having it in a single folder is necessary
for us to upload the information using the VSTS Build Artifact task.

Jenkins should be unaffected. It supports and uses wildcards for the
file types we need to be uploaded on failed builds.