Downgrade desktop projects to net46
Even though Roslyn can only every be deployed where at least net461 or netcoreapp2.0 is available Roslyn cannot depend on netstandard2.0 assemblies. This is due to the setup of Visual Studio which has a minimum target of net46 and hence only deploys netstandard1.3 facades. In order to prevent netstandard2.0 versions of NuGet references to deploy, at build or test, we must limit the desktop target framework used in our projects to net46.

There is nothing functionally wrong with having our tests target net461. At the same time though it would mean our tests were running against the netstandard2.0 version of references, such as System.Collections.Immutable, which is not what we deploy against. Hence not testing what we ship which creates a small but real hole.