Clean up dependencies projects (#25609)
* Remove src/Setup/DevDivPackagesRoslyn/

This contained a list of packages that we depended on and needed
facades deployed into Visual Studio. We already have
VisualStudioSetup.Dependencies that already (should have...) contained
all the facades and more. We can just consume that project instead
of having yet another duplication of facade lists.

* Fix casing

* Also include the CompilerExtension as a source of facades that we must include

This ensures that if the compiler adds a new facade but we don't update
VisualStudioSetup.Dependencies (either intentionally or by accident)
we still will include the right facades.

* Include DevDivPackagesDebugger into Roslyn.sln

By doing this, the VisualStudioSetup.Dependencies.csproj will now
figure out the highest version of our own dependencies and the debugger
dependencies, and create a single package that we can deploy that
contains all of them.

* Exclude inserting packages that are created by VS itself

* Remove ProjectGuid from DevDivPackagesDebugger.csproj

This makes BuildBoss happy.