Add per-syntax tree diagnostic options to CompilationOptions
This feature is necessary for editorconfig, but disconnected from it
directly. The idea is that you can manipulate diagnostics on a
per-syntax tree basis, not just at the compilation level. The design is
that the more specific diagnostic configuration wins. The ordering of
specifity, from least to most specific, is as follows:

1. Compilation-level.
2. Syntax tree-level.
3. Pragma-level.

Thus, if a diagnostic is configured in contradictory ways at the
syntax tree level and the compilation level, the setting at the syntax
tree level will win.