Report an error when discard pattern is used but there is a _ in scope Note that this error does not apply to the discard *designation*, only the discard *pattern*. Since the discard designation is new in C#8, this is not a breaking change.
Report an error if a constant named '_' is used as a constant pattern.
This error has been confirmed by the LDM on 2017-11-20

Also report a warning if a type named '_' is used in an is-type expression.
This requires compat council confirmation.

Open issues in https://github.com/dotnet/csharplang/issues/1054 describe pending LDM decisions for potentialy breaking scenarios.