Add diagnostic options to SyntaxTree
This change allows the user of a Compilation to configure warning
reporting on a syntax tree level. The design is that the more specific
diagnostic configuration wins. The ordering of diagnostic specifity,
from least to most specific, is as follows:

    1. Compilation level
    2. Syntax tree level
    3. Pragma level

If the diagnostic options at two levels are contradictory, the more
specific setting wins.