Avoid scheduling work that's already cancelled
While it doesn't make sense to still be requesting foreground actions after
cancellation is requested, tracking down all callers of
IForegroundNotificationService to ensure the proper checks are in place can be
tedious. This change treats IForegroundNotificationService as the gatekeeper
for scheduling asynchronous operations, and proactively cancels operations if
requested before adding them to the work queue.