Don't include AnalyzerFileReference.Display in equality checking
We were including AnalyzerFileReference.Display as a part of equality
checking in AnalyzerFileReference. This is expensive to acquire, as it
hits the disk and loads the assembly image to read the assembly name
and identity from metadata. This isn't necessary to do: since we were
already using FullPath as equality checking, the only way this would
ever matter is if you had two AnalyzerFileReferences that were created
at different times, with the assembly being modified in between. Since
AnalyzerFileReference still had no concept of snapshotting, the behavior
was still undefined. According to @heejaechang the analyzer runner
already filters out "identical" analyzers with the same .NET assembly
identity (since have no way to load both anyways...) so this was even
more silly.

This was showing up in various traces, as the workspace has various
"does this AnalyzerReference exist in this project", which was hitting
disk (often on the UI thread) when it wasn't needed.