Properly type forward Microsoft.CodeAnalysis.FileTextLoader
Since we're moving this type from Microsoft.CodeAnalysis.Workspaces.Desktop to Microsoft.CodeAnalysis.Workspaces, a type forward is
required. There is currently a bug with the Public API analyzer with properties on forwarded types that I have to work around in the
PublicAPI.Shipped.txt, but a fix is already out for that https://github.com/dotnet/roslyn-analyzers/pull/1657. Once the analyzers are
updated in the roslyn repo, this will get fixed.