Mark lambda conversions as side-effecting
The scenario listed in https://github.com/dotnet/roslyn/issues/22700
demonstrates the problem. The direct cause of the crash is that local
lowering produces two instances of the lambda and lamda conversions in
the resulting bound nodes. This crashes the closure conversion pass
since the same lambda cannot appear in two places in the tree. However,
even if that were allowed, this would violate the language
specification.

Placing a lambda inside an indexer which is nested in a compound
assignment creates two calls which contain the delegate produced by the
lambda expression. If the expression kind is marked as side-affecting it
will be spilled into a temporary variable. However, the local lowering
doesn't distinguish between whether an expression is side-affecting and
whether or not it is referentially transparent. In this case, both
matter. Since the delegate produced by the conversion can be compared by
reference to any other delegate produced by the same lambda, the code
must be referentially transparent -- it must produce the same result
when evaluated repeatedly.

The fix is small: mark lambda conversions as side-affecting. There are
certain cases, like the creation of temporaries for out-of-order calls
with named parameters, where the lambda conversion is only required to
be non-side-affecting instead of referentially transparent, but it does
not currently seem worth the implementation effort to split up the code
path.

Fixes #22700